+++ Sources/SwiftiomaticKit/Configuration/Configuration.swift
@@ -414,52 +414,5 @@ extension Configuration: Codable {
         }
     }
 
-    private enum JSONValue: Codable {
-        case string(String)
-        case int(Int)
-        case double(Double)
-        case bool(Bool)
-        case array([JSONValue])
-        case object([String: JSONValue])
-        case null
-
-        init(from decoder: any Decoder) throws {
-            let container = try decoder.singleValueContainer()
-            if let v = try? container.decode(Bool.self) {
-                self = .bool(v)
-            } else if let v = try? container.decode(Int.self) {
-                self = .int(v)
-            } else if let v = try? container.decode(Double.self) {
-                self = .double(v)
-            } else if let v = try? container.decode(String.self) {
-                self = .string(v)
-            } else if let v = try? container.decode([JSONValue].self) {
-                self = .array(v)
-            } else if let v = try? container.decode([String: JSONValue].self) {
-                self = .object(v)
-            } else if container.decodeNil() {
-                self = .null
-            } else {
-                throw DecodingError.dataCorrupted(
-                    .init(
-                        codingPath: decoder.codingPath,
-                        debugDescription: "Unsupported JSON value"
-                    )
-                )
-            }
-        }
-
-        func encode(to encoder: any Encoder) throws {
-            var container = encoder.singleValueContainer()
-            switch self {
-            case .string(let v): try container.encode(v)
-            case .int(let v): try container.encode(v)
-            case .double(let v): try container.encode(v)
-            case .bool(let v): try container.encode(v)
-            case .array(let v): try container.encode(v)
-            case .object(let v): try container.encode(v)
-            case .null: try container.encodeNil()
-            }
-        }
-    }
 }
+